<fix>[data]: Support aarch64 architecture strongSwan upgrade
Resolves: ZSTAC-48602

+++ data/upgrade/strongswan/5.9.4/upgrade.sh
@@ -1,11 +1,14 @@
 #!/bin/bash
 
-if [ `grep -c "1.1.7" /opt/vyatta/etc/version` -eq '0' ]; then
+set -ex
+
+if [ `grep -c "1.1.7" /opt/vyatta/etc/version` -eq '0' ] && [ `grep -c "1.2" /opt/vyatta/etc/version` -eq '0' ]; then
     exit 0
 fi
 
 DATA_PATH="/home/vyos/zvr/data"
 SW_594_PATH="${DATA_PATH}/upgrade/strongswan/5.9.4"
+CURRENT_ARCH=`uname -m`
 
 case "$1" in
     "-d")
@@ -17,16 +20,22 @@ case "$1" in
         /usr/bin/dpkg -r strongswan-zstack
         ;;
     *)
-        /usr/bin/dpkg -i ${SW_594_PATH}/strongswan-zstack_5.9.4-1_amd64.deb
+        if [[ x"$CURRENT_ARCH" == x"x86_64" ]]; then
+            /usr/bin/dpkg -i ${SW_594_PATH}/strongswan-zstack_5.9.4-1_amd64.deb
+        elif [[ x"$CURRENT_ARCH" == x"aarch64" ]];then
+            /usr/bin/dpkg -i ${SW_594_PATH}/strongswan-zstack_5.9.4-1_arm64.deb
+        fi
         cp ${SW_594_PATH}/ipsec.conf /usr/local/etc/ipsec.conf
         cp ${SW_594_PATH}/ipsec.secrets /usr/local/etc/ipsec.secrets
         cp ${SW_594_PATH}/strongswan.conf /usr/local/etc/strongswan.conf
 
-        libcrypto_file=/usr/lib/libcrypto.so.1.0.0
-        if [[ -L "$libcrypto_file" ]]; then
-            rm -rf $libcrypto_file 2>/dev/null
+        if [[ x"$CURRENT_ARCH" == x"x86_64" ]]; then
+            libcrypto_file=/usr/lib/libcrypto.so.1.0.0
+            if [[ -L "$libcrypto_file" ]]; then
+                rm -rf $libcrypto_file 2>/dev/null
+            fi
+            ln -s /usr/local/lib/libcrypto.so.1.0.0 $libcrypto_file
         fi
-        ln -s /usr/local/lib/libcrypto.so.1.0.0 $libcrypto_file
         ;;
 esac
